chore: move mainApp to different file
